+++ src/icalendar/cal.py
@@ -141,14 +141,19 @@ class Component(CaselessDict):
         if isinstance(value, types_factory.all_types):
             # Don't encode already encoded values.
             return value
-        klass = types_factory.for_property(name)
-        obj = klass(value)
         if parameters:
             if isinstance(parameters, dict):
                 params = Parameters()
                 for key, item in parameters.items():
                     params[key] = item
                 parameters = params
+        klass = types_factory.for_property(
+            name, parameters.get('VALUE') if parameters else None)
+        if types_factory.is_list_property(name):
+            obj = vDDDLists(value, klass)
+        else:
+            obj = klass(value)
+        if parameters:
             assert isinstance(parameters, Parameters)
             obj.params = parameters
         return obj
@@ -185,7 +190,7 @@ class Component(CaselessDict):
 
         # encode value
         if encode and isinstance(value, list) \
-                and name.lower() not in ['rdate', 'exdate']:
+                and not types_factory.is_list_property(name):
             # Individually convert each value to an ical type except rdate and
             # exdate, where lists of dates might be passed to vDDDLists.
             value = [self._encode(name, v, parameters, encode) for v in value]
@@ -217,7 +222,11 @@ class Component(CaselessDict):
         if isinstance(value, vDDDLists):
             # TODO: Workaround unfinished decoding
             return value
-        decoded = types_factory.from_ical(name, value)
+        try:
+            valtype = value.params['VALUE']
+        except (AttributeError, KeyError):
+            valtype = None
+        decoded = types_factory.from_ical(name, value, valtype)
         # TODO: remove when proper decoded is implemented in every prop.* class
         # Workaround to decode vText properly
         if isinstance(decoded, vText):
@@ -225,11 +234,8 @@ class Component(CaselessDict):
         return decoded
 
     def decoded(self, name, default=_marker):
-        """Returns decoded value of property.
+        """Returns value of a property as a python native type.
         """
-        # XXX: fail. what's this function supposed to do in the end?
-        # -rnix
-
         if name in self:
             value = self[name]
             if isinstance(value, list):
@@ -369,7 +375,6 @@ class Component(CaselessDict):
                     _timezone_cache[component['TZID']] = component.to_tz()
             # we are adding properties to the current top of the stack
             else:
-                factory = types_factory.for_property(name)
                 component = stack[-1] if stack else None
                 if not component:
                     raise ValueError('Property "{prop}" does not have '
@@ -377,10 +382,18 @@ class Component(CaselessDict):
                 datetime_names = ('DTSTART', 'DTEND', 'RECURRENCE-ID', 'DUE',
                                   'FREEBUSY', 'RDATE', 'EXDATE')
                 try:
-                    if name in datetime_names and 'TZID' in params:
-                        vals = factory(factory.from_ical(vals, params['TZID']))
+                    factory = types_factory.for_property(name,
+                                                         params.get('VALUE'))
+                    if types_factory.is_list_property(name):
+                        vals = vDDDLists(
+                            vDDDLists.from_ical(vals, params.get('TZID'),
+                                                factory))
                     else:
-                        vals = factory(factory.from_ical(vals))
+                        if name in datetime_names and 'TZID' in params:
+                            vals = factory(
+                                factory.from_ical(vals, params['TZID']))
+                        else:
+                            vals = factory(factory.from_ical(vals))
                 except ValueError as e:
                     if not component.ignore_exceptions:
                         raise